Remove autogenerated Android Studio files
Although it is recommended to keep them under source control, these
files can contain a lot of machine-specific information, like full
absolute paths, JVM version etc. This is unacceptable for a shared
project. IDE generates a lot of unneccessary changes each time the
project moves from one machine to another. In the best case, it's just
extra noise in repository. In the worst case, this leads to conflicts.